Reflection implementation + many other things
-> Attempted to implement Reflection using polymorphism, concepts and type punning.
-> Written equality operator for mathematical types( Vectors, Quaternions...)
-> Fixed some assigment operators for the mathematical types
-> Removed Engine3DRadSpace::ComplexNumber, it can be replaced by std::complex<float>

-> Fixed few warnings related to unsigned and size_t overflows

There are test cases made in a separate VS project, using Google Tests:

-> Written few test cases for the mathematical types (LinearAlgebraTests)
-> Written some test cases for reflection (ReflectionTests)

+++ 3DRadSpace/3DRadSpaceDll/Model3D.cpp
@@ -151,7 +151,7 @@ Engine3DRadSpace::Model3D::Model3D(Game* game, const char* file)
 		}
 
 		//create index buffer
-		size_t indexbuffersize = scene->mMeshes[i]->mNumFaces * 3; //faces are triangulated, so multiply with 3
+		size_t indexbuffersize = static_cast<size_t>(scene->mMeshes[i]->mNumFaces) * 3; //faces are triangulated, so multiply with 3
 
 		unsigned* indexbuffer = new unsigned[indexbuffersize];
